fix(memory-wiki): refresh compiled dashboard state
